Refactor forward console test
