+++ tests/console/config/local.json
@@ -1,0 +1,1 @@
+{}
