+++ tests/console/run
@@ -1,0 +1,8 @@
+#!/usr/bin/env node
+
+const pentf = require('../../index.js');
+
+pentf.main({
+    rootDir: __dirname,
+    testsDir: __dirname,
+});
